refactor: moved clock types to session
